requirements.txt
discord.py>=2.5.2
# python-dotenv>=1.1.0  # Use for local testing (without Docker)
aiohttp>=3.11.16